ARCH-3: Extracted models into separate directory and created mocked data lists for each model
