+++ WineUp/WineUp/Models/CatalogFiltersBarItemModel.swift
@@ -1,0 +1,13 @@
+//
+//  CatalogFiltersBarItemModel.swift
+//  WineUp
+//
+//  Created by Александр Пахомов on 04.10.2020.
+//
+
+import Foundation
+
+struct CatalogFiltersBarItemModel: Identifiable {
+    var id = UUID()
+    var title: String
+}
